Update README.md
In this article proposed a novel feature descriptor for estimating the kinship relationship between a parent and their child. Facial Kinship Verification is a challenging research problem due to the lack of significant facial similarities between parents and children.
The proposed approach addresses this issue by utilizing childhood images of the parent to increase the facial resemblance with their child. The facial descriptors are extracted from sub-bands of curvelet transform to capture the distinctive features of
facial images. In addition, texture information is extracted using a convolutional neural network. Statistical measures are employed to assess the similarity between the images, and a threshold is calculated during the training phase to classify the kinship
relationship in the test phase. The proposed method was evaluated on several databases and demonstrated superior performance compared to existing state-of-the-
art methods.